Version 0.4.0: MSRV bump, soundness fixes
This is a breaking release because the MSRV has been bumped to an
extremely recent version of Rust (1.63.0). The reason for this is that
`NonNull::len` has finally stabilized, allowing `imgref-iter` to read
slice-ptr lengths without upsetting Miri.

It's not entirely known if the older behavior was UB. I will not be
yanking the older versions, but I strongly recommend upgrading.